refactor numeric types
