+++ frontends/systolic-lang/convert-mems.py
@@ -1,6 +1,6 @@
 import argparse
 import json
-from fud.stages.verilator import numeric_types
+from calyx import numeric_types
 
 
 if __name__ == "__main__":
