+++ fud/fud/stages/verilator/json_to_dat.py
@@ -1,6 +1,6 @@
 import simplejson as sjson
 import numpy as np
-from .numeric_types import FixedPoint, Bitnum
+from calyx.numeric_types import FixedPoint, Bitnum
 from pathlib import Path
 from fud.errors import InvalidNumericType, Malformed
 import logging as log
